+++ adapters/AppLovin/AppLovinAdapter/GADMAdapterAppLovinInterstitialDelegate.m
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 #import "GADMAdapterAppLovinInterstitialDelegate.h"
+#import "AppLovinAdapter-Swift.h"
 #import "GADMAdapterAppLovinMediationManager.h"
-#import "GADMAdapterAppLovinUtils.h"
 
 @implementation GADMAdapterAppLovinInterstitialDelegate {
   /// AppLovin interstitial ad renderer to which the events are delegated.
@@ -36,13 +36,13 @@
 - (void)adService:(nonnull ALAdService *)adService didLoadAd:(nonnull ALAd *)ad {
   GADMAdapterAppLovin *parentRenderer = _parentRenderer;
 
-  BOOL isMultipleAdsEnabled = GADMAdapterAppLovinIsMultipleAdsLoadingEnabled();
+  BOOL isMultipleAdsEnabled = [GADMAdapterAppLovinUtils isMultipleAdsLoadingEnabled];
   if (isMultipleAdsEnabled) {
     [GADMAdapterAppLovinMediationManager.sharedInstance
         removeInterstitialZoneIdentifier:parentRenderer.zoneIdentifier];
   }
 
-  [GADMAdapterAppLovinUtils log:@"Interstitial did load ad: %@", ad];
+  [GADMAdapterAppLovinUtils log:[NSString stringWithFormat:@"Interstitial did load ad: %@", ad]];
   parentRenderer.interstitialAd = ad;
   [parentRenderer.connector adapterDidReceiveInterstitial:parentRenderer];
 }
@@ -51,7 +51,7 @@
   GADMAdapterAppLovin *parentRenderer = _parentRenderer;
   [GADMAdapterAppLovinMediationManager.sharedInstance
       removeInterstitialZoneIdentifier:parentRenderer.zoneIdentifier];
-  NSError *error = GADMAdapterAppLovinSDKErrorWithCode(code);
+  NSError *error = [GADMAdapterAppLovinUtils sdkErrorWithCode:code];
   [parentRenderer.connector adapter:parentRenderer didFailAd:error];
 }
 
@@ -91,8 +91,10 @@
 - (void)videoPlaybackEndedInAd:(nonnull ALAd *)ad
              atPlaybackPercent:(nonnull NSNumber *)percentPlayed
                   fullyWatched:(BOOL)wasFullyWatched {
-  [GADMAdapterAppLovinUtils log:@"Interstitial video playback ended at playback percent: %lu%%",
-                                (unsigned long)percentPlayed.unsignedIntegerValue];
+  [GADMAdapterAppLovinUtils
+      log:[NSString
+              stringWithFormat:@"Interstitial video playback ended at playback percent: %lu%%",
+                               (unsigned long)percentPlayed.unsignedIntegerValue]];
 }
 
 @end