+++ adapters/AppLovin/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## AppLovin iOS Mediation Adapter Changelog
 
+#### Next version
+- Migrated GADMAdapterAppLovinUtils to Swift.
+
 #### [Version 13.6.3.0](https://dl.google.com/googleadmobadssdk/mediation/ios/applovin/AppLovinAdapter-13.6.3.0.zip)
 - Verified compatibility with AppLovin SDK 13.6.3.
 - Update the minimum iOS version to relax Google Mobile Ads SDK dependency to allow minor versions.